move userOnlineMinutes to instance settings
