+++ scripts/core/instance-settings-interface.ts
@@ -8,21 +8,10 @@
 // !!!
 
 export interface IInstanceSettings {
-    authoring: {
-        editor: {
-            /**
-             * Version 2 is deprecated.
-             */
-            version: '2' | '3';
-        };
-        customToolbar: boolean;
-    };
-
-    monitoring: {
-        listOfStrings: Array<string>;
-        listOfNumbers: Array<number>;
-        listOfBooleans: Array<boolean>;
-        listOfUnions: Array<'item 1' | 'item 2' | 'item 3'>;
-        listOfObjects: Array<{firstName: string; age: number}>;
+    users: {
+        /**
+         * Time of inactivity in minutes until user is no longer considered online.
+         */
+        minutesOnline: number;
     };
 }
