+++ scripts/index.ts
@@ -1,14 +1,5 @@
-import {IInstanceSettings} from 'core/instance-settings-interface';
-import {PartialDeep} from 'core/helpers/typescript-helpers';
 import {defaultsDeep} from 'lodash';
-
-const instanceSettings: IInstanceSettings = {} as IInstanceSettings; // will be populated in loadInstanceSettings
-
-const defaultInstanceSettings: PartialDeep<IInstanceSettings> = {
-    authoring: {
-        customToolbar: false,
-    },
-};
+import {instanceSettings, defaultInstanceSettings} from 'instance-settings';
 
 function loadInstanceSettings() {
     const xhr = new XMLHttpRequest();
@@ -25,9 +16,6 @@ function loadInstanceSettings() {
     });
 
     xhr.open('GET', `${__SUPERDESK_CONFIG__.server.url}/config/instance-settings`, asynchronous);
-
-    // TODO: it should be possible to get instance settings without having to log in
-    xhr.setRequestHeader('Authorization', 'Basic OWM3OTRmN2UtZGEzYy00MzNkLWJjZGQtOTdjNWRiNjdiYjFiOg==');
 
     xhr.send();
 }
