+++ webpack.config.js
@@ -277,7 +277,5 @@ function getDefaults(grunt) {
             'pt_BR',
             'pl',
         ],
-
-        userOnlineMinutes: 15,
     };
 }
